remove resolveToCallExpression and merge conditions to resolveToValue
